+++ src/lib/types/index.ts
@@ -10,4 +10,5 @@ export * from "./projects";
 export * from "./addrs";
 export * from "./proposal";
 export * from "./validator";
+export * from "./block";
 export * from "./asset";