+++ CHANGELOG.md
@@ -39,6 +39,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- [#286](https://github.com/alleslabs/celatone-frontend/pull/286) Add block proposer
+- [#275](https://github.com/alleslabs/celatone-frontend/pull/275) Searching can now be determined by chain
+- [#273](https://github.com/alleslabs/celatone-frontend/pull/273) Add subheader and network overview
+- [#278](https://github.com/alleslabs/celatone-frontend/pull/278) Fully wired block details page
+- [#272](https://github.com/alleslabs/celatone-frontend/pull/272) Add blocks page and table component
 - [#287](https://github.com/alleslabs/celatone-frontend/pull/287) Add app default human address for each chain
 - [#271](https://github.com/alleslabs/celatone-frontend/pull/271) Create transaction list page
 - [#245](https://github.com/alleslabs/celatone-frontend/pull/245) Add amplitude for account detail page
@@ -57,6 +62,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- [#293](https://github.com/alleslabs/celatone-frontend/pull/293) Add comma separator to pagination and total blocks
+- [#291](https://github.com/alleslabs/celatone-frontend/pull/291) Update tx count query
+- [#285](https://github.com/alleslabs/celatone-frontend/pull/285) Add data fetching error state, and empty state to blocks page
 - [#292](https://github.com/alleslabs/celatone-frontend/pull/292) Handle format token with M and show only 2 decimal points when more than 1,000
 - [#284](https://github.com/alleslabs/celatone-frontend/pull/284) Add data fetching error state to txs page
 - [#281](https://github.com/alleslabs/celatone-frontend/pull/281) Change Osmosis mainnet GraphQL